[refactor]Layout 오류 수정
